requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
da-vinci==0.3.0
datetime-truncate==1.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-admin-charts==1.0.0
django-admin-list-charts==0.1.3
django-admincharts==0.4.0
django-bower==5.2.0
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-thumbnails==0.7.0
django-timezone-field==5.0
django-user-agents==0.4.0
djangorestframework==3.13.1
djcacheutils==3.0.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
pycparser==2.21
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
redis==4.3.4
segno==1.5.2
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
ua-parser==0.15.0
Unidecode==1.3.4
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
